Перерабатывает и выносит функцию масштабирования фотографий в отдельный модуль. Добавляет экспорт этой фукнции в image-upload-form.js.
